Test JDBC flush_metadata_cache sequentially
Extract JDBC's `flush_metadata_cache` procedure tests into separate
class and run it sequentially.  These tests interfered with each other
when running in parallel.
